Refactoring backend-core tenancy to context, splitting out tenancy functionality to its own area.
